XML Reader Changes
I made changes to the Team LIst window.  I believe it will now work if
we dispose and reopen the window.  There is also no more need for a
file to keep track of what files we have.  Changes I made to
AddTeamWindow and GUI were minor but nessesary.  No more need to save
and load, only in the constructor of TeamListWindow.